requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
chardet==3.0.4
charset-normalizer==3.1.0
discord.py==2.2.2
docopt==0.6.2
filelock==3.10.7
frozenlist==1.3.3
googletrans==3.1.0a0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.13.3
hyperframe==5.2.0
idna==2.10
multidict==6.0.4
numpy==1.24.2
packaging==23.0
pipreqs==0.4.11
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
rfc3986==1.5.0
sniffio==1.3.0
tokenizers==0.13.2
tqdm==4.65.0
transformers==4.27.3
typing-extensions==4.5.0
urllib3==1.26.15
yarg==0.1.9
yarl==1.8.2
